emotion-reco-api/pyproject.toml
[tool.poetry]
name = "emotion-reco-api"
version = "0.1.0"
description = ""
authors = ["zsalazarq <salazarsergio1082@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11"
django = "^5.2.6"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.16.1"
django-filter = "^25.1"
python-dotenv = "^1.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
